+++ data/src/main/java/br/com/viniciuscoscia/data/remote/model/route/RouteRequest.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 
 data class RouteRequest(
         @SerializedName("places")
-    val places: List<Place> = listOf(),
+        val places: List<Place> = listOf(),
         @SerializedName("fuel_consumption")
-        val fuelConsumption: Double = 0.0,
+        val fuelConsumption: Int = 0,
         @SerializedName("fuel_price")
-    val fuelPrice: Double = 0.0
+        val fuelPrice: Double = 0.0
 )
